Waking animation works semi seamlessly
