+++ SmartButler/Assets/Scripts/AnimatedCamera.cs.meta
