+++ SmartButler/Assets/Scripts/AutoFade.cs
