+++ SmartButler/Assets/Scripts/AutoFade.cs.meta
