+++ SmartButler/Assets/Scripts/GameManager/GameStateManager.cs
@@ -108,6 +108,8 @@ namespace Assets.Scripts.GameManager {
                     GameState = State.Stereo;
                     goto default;
                 default:
+                    if (!_LivingRoomsSoundtrack)
+                        _LivingRoomsSoundtrack = FindObjectOfType<LivingRoomSoundtrack>();
                     switch (GameState) {
 						case State.Clock:
 							_clockLoader.IsEnterable = false;
