+++ SmartButler/Assets/Scripts/Lights.cs
